app/file_utils.py
"""
Resume File Processing Module

This module provides utilities for handling and processing resume file uploads in the
ElevatrAI platform. It supports extracting text content from PDF and Word documents
(DOC/DOCX) while maintaining memory efficiency and security.

Key Features:
- Memory-efficient file processing using BytesIO
- Support for PDF and Word document formats
- Secure file handling with format validation
- Clean error handling with descriptive messages

Dependencies:
    - PyPDF2: For PDF file processing
    - python-docx: For Word document processing
    - werkzeug: For secure file handling

Author: Anslem Akadu
"""
import io
import PyPDF2
import docx
from werkzeug.datastructures import FileStorage

def process_resume_upload(file: FileStorage) -> str:
    """
    Extract text content from an uploaded resume file safely and efficiently.
    
    This function processes resume uploads by:
    1. Validating the file format (PDF/DOC/DOCX)
    2. Reading the file content into memory safely
    3. Extracting text based on file type
    4. Performing basic content validation
    
    Args:
        file (FileStorage): The uploaded file object from a Flask request
        
    Returns:
        str: Extracted text content from the resume
        
    Raises:
        ValueError: If the file is missing, format is unsupported, or text extraction fails
        
    Example:
        from flask import request
        
        @app.route('/upload', methods=['POST'])
        def handle_upload():
            file = request.files['resume']
            try:
                text = process_resume_upload(file)
                # Process the extracted text...
            except ValueError as e:
                return str(e), 400
    """
    # Input validation
    if not file or not file.filename:
        raise ValueError("No file provided")
        
    # Read file into memory once for efficiency
    file_content = file.read()
    filename = file.filename.lower()
    
    try:
        # Process based on file type
        if filename.endswith('.pdf'):
            # PDF Processing: Use BytesIO for memory efficiency
            pdf_file = io.BytesIO(file_content)
            reader = PyPDF2.PdfReader(pdf_file)
            text = ' '.join(page.extract_text() for page in reader.pages)
            
        elif filename.endswith(('.doc', '.docx')):
            # Word Doc Processing: Handle both DOC and DOCX
            doc_file = io.BytesIO(file_content)
            doc = docx.Document(doc_file)
            text = ' '.join(paragraph.text for paragraph in doc.paragraphs)
            
        else:
            raise ValueError("Unsupported file format. Please upload PDF or Word document.")
        
        # Validate extracted content
        if not text.strip():
            raise ValueError("No text content found in file")
            
        return text
            
    except Exception as e:
        # Provide helpful error message while hiding implementation details
        raise ValueError(f"Failed to extract text from file: {str(e)}")
        
    finally:
        # Reset file pointer to allow for potential reuse
        file.seek(0)
        
# TODO: Add support for more document formats (RTF, TXT)
# TODO: Add content sanitization
# TODO: Add advanced text extraction (headers, sections, etc.)
# TODO: Add support for parsing structured resume sections
