requirements.txt
flask==2.3.3
Werkzeug==2.3.7
sentence-transformers==2.2.2
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==21.2.0
torch==2.0.1
python-dotenv==1.0.0
